% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_modules_expected}
\alias{check_modules_expected}
\title{Check the marks folder contents against file "modules_expected_here.txt"}
\usage{
check_modules_expected()
}
\value{
A list with two character components, named
\code{extras} and \code{missing}
}
\description{
Check the marks folder contents against file "modules_expected_here.txt"
}
