% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_module_effects}
\alias{get_module_effects}
\title{Print and return a table of module effects}
\usage{
get_module_effects(module_codes, mdd)
}
\arguments{
\item{module_codes}{Character; the vector of module codes used in the report}

\item{mdd}{List of median differences}
}
\value{
A 2-column data frame with columns \code{Effect} and \code{Count}
}
\description{
Print and return a table of module effects
}
