% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{history_boxplot}
\alias{history_boxplot}
\title{Boxplots of recent marks data}
\usage{
history_boxplot(history)
}
\arguments{
\item{history}{A data frame containing marks history, as created by [norman::save_history()].}
}
\description{
These are based on summary statistics from the `history.csv` file in the working folder (if there is one),
filtered on the five most recent years. Note that because there are based on the minimum, maximum, and quartiles of the marks data,
rather than the raw marks, outliers aren't shown.
}
\seealso{
[norman::n_recent_years()], [norman::save_history()]
}
