% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{history_effects}
\alias{history_effects}
\title{Line graph of recent module effects}
\usage{
history_effects(history)
}
\arguments{
\item{history}{A data frame containing marks history, as created by [norman::save_history()].}
}
\description{
Created from the module effects saved in `history.csv`
}
\seealso{
[norman::n_recent_years()], [norman::save_history()]
}
