% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meddiff}
\alias{meddiff}
\title{Compute a matrix of median differences}
\usage{
meddiff(xmat)
}
\arguments{
\item{xmat}{A numeric matrix}
}
\value{
A square numeric matrix, with size equal to the number of
columns in \code{xmat}
}
\description{
Compute a matrix of median differences
}
