% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meddiff_fit}
\alias{meddiff_fit}
\title{Extract module effects from the median differences}
\usage{
meddiff_fit(m)
}
\arguments{
\item{m}{A numeric matrix of mdeian differences as computed by \code{meddiff}}
}
\value{
A \code{lm} model object
}
\description{
Extract module effects from the median differences
}
