% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{n_recent_years}
\alias{n_recent_years}
\title{Get the most recent years of data
Assumes the df being filtered was created by `save_history()`}
\usage{
n_recent_years(history, n = 5)
}
\arguments{
\item{history}{A data frame containing marks history, as created by [norman::save_history()].}

\item{n}{integer - the number of most recent years}
}
\description{
Get the most recent years of data
Assumes the df being filtered was created by `save_history()`
}
