% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_file_listing}
\alias{print_file_listing}
\title{Print and return the list of files in a folder}
\usage{
print_file_listing(folder)
}
\arguments{
\item{folder}{Character, the relative path to a folder from the
current working directory}
}
\value{
Character; a vector of file names
}
\description{
Print and return the list of files in a folder
}
