% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_module_effect}
\alias{print_module_effect}
\title{Get the module effect for a given module code}
\usage{
print_module_effect(module_code)
}
\arguments{
\item{module_code}{Character; the length-5 module code}
}
\value{
Character; the module effect, rounded to 1 decimal place.
}
\description{
Get the module effect for a given module code
}
