% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_path}
\alias{print_path}
\title{Get the full path of a folder}
\usage{
print_path(folder)
}
\arguments{
\item{folder}{Character, the relative path to a folder from the
current working directory}
}
\value{
Character, the full path to the folder.  (invisibly)
}
\description{
Get the full path of a folder
}
