% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raw_mark_classes}
\alias{raw_mark_classes}
\title{Make a matrix of percentages in each degree class range, for all modules}
\usage{
raw_mark_classes(marks_matrix, dp = 0)
}
\arguments{
\item{marks_matrix}{The full matrix of marks for the modules}

\item{dp}{Numeric, the number of decimal places to round the result to}
}
\value{
A matrix
}
\description{
Make a matrix of percentages in each degree class range, for all modules
}
