% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raw_mark_summaries}
\alias{raw_mark_summaries}
\title{Make a matrix of 8-number summaries for all modules}
\usage{
raw_mark_summaries(marks_matrix)
}
\arguments{
\item{marks_matrix}{The full matrix of marks for the modules}
}
\value{
A matrix
}
\description{
Make a matrix of 8-number summaries for all modules
}
