% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_history}
\alias{save_history}
\title{Save history}
\usage{
save_history()
}
\arguments{
\item{year}{The year}
}
\description{
If a "history.csv" file exists, adds the eight-number summary and module
effect to that file, one row per module.
If "history.csv" does not exist, creates the file first.
}
\details{
history.csv must have the following columns:

- Year
- Module
- (N)
- Zeros
- Min.
- 1st Qu.
- Median
- 3rd Qu.
- Max.
- Mean,
- S.D.
}
\examples{
\dontrun{
norman::save_history(2025)
}
}
