% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatter}
\alias{scatter}
\title{Make a module-specific scatterplot}
\usage{
scatter(module_code)
}
\arguments{
\item{module_code}{Character; the length-5 module code}
}
\value{
A \code{ggplot} object
}
\description{
Make a module-specific scatterplot
}
