% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stemleaf}
\alias{stemleaf}
\title{Make a module-specific stem and leaf plot}
\usage{
stemleaf(module_code, module_marks)
}
\arguments{
\item{module_code}{Character; the length-5 module code}

\item{module_marks}{List of module marks, one component per module}
}
\value{
A character vector for display
}
\description{
Make a module-specific stem and leaf plot
}
