% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update}
\alias{update}
\title{Update the \code{norman} package --- a wrapper for \code{remotes::install_github}}
\usage{
update(build_opts = "--no-build-vignettes", ...)
}
\arguments{
\item{build_opts}{Character; options for \code{R CMD build}.  Default is \code{"--no-build-vignettes"}.}

\item{...}{Other arguments to pass to \code{remotes::install_github}}
}
\description{
Update the \code{norman} package --- a wrapper for \code{remotes::install_github}
}
\examples{
\dontrun{
norman::update()
norman::update(force = TRUE)
}

}
