% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_module_effects}
\alias{get_module_effects}
\title{Print and return a table of module effects}
\usage{
get_module_effects()
}
\value{
A 2-column data frame with columns \code{Effect} and \code{Count}
}
\description{
Print and return a table of module effects
}
