% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_all_median_differences}
\alias{list_all_median_differences}
\title{List all median within-student differences between modules}
\usage{
list_all_median_differences(mdd)
}
\arguments{
\item{mdd}{A list}
}
\value{
\code{invisible(NULL)}
}
\description{
List all median within-student differences between modules
}
\examples{
#
# Toy example from
# https://davidfirth.github.io/blog/2019/04/26/robust-measurement-from-a-2-way-table/
#
x <- structure(c(NA, NA, 10, NA, NA, 20, NA, NA, 30, 45, 55, NA, 60, 60, 50),
  .Dim = c(3L, 5L), .Dimnames = structure(list(student = c("i", "j", "k"),
  module = c("A", "B", "C", "D", "E")), .Names = c("student", "module")))
print(x)
mdd <- meddiff_for_display(x)
list_all_median_differences(mdd)

}
