% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_module_pages}
\alias{make_module_pages}
\title{Make a report page for each module}
\usage{
make_module_pages(keep_tmpdir = FALSE)
}
\arguments{
\item{keep_tmpdir}{Logical; whether to keep the working "tmp" directory}
}
\value{
Character; R Markdown text for the module pages that were made
}
\description{
Make a report page for each module
}
