% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meddiff}
\alias{meddiff}
\title{Compute a matrix of median differences}
\usage{
meddiff(xmat)
}
\arguments{
\item{xmat}{A numeric matrix}
}
\value{
A square numeric matrix, with size equal to the number of
columns in \code{xmat}
}
\description{
Compute a matrix of median differences
}
\examples{
#
# Toy example from
# https://statgeek.net/2019/04/26/robust-measurement-from-a-2-way-table
#
x <- structure(c(NA, NA, 10, NA, NA, 20, NA, NA, 30, 45, 55, NA, 60, 60, 50),
  .Dim = c(3L, 5L), .Dimnames = structure(list(student = c("i", "j", "k"),
  module = c("A", "B", "C", "D", "E")), .Names = c("student", "module")))
print(x)
meddiff(x)

}
