% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meddiff_for_display}
\alias{meddiff_for_display}
\title{A version of \code{meddiff} to compute median differences in a different format}
\usage{
meddiff_for_display(xmat)
}
\arguments{
\item{xmat}{A numeric matrix}
}
\value{
A list, with one vector component for each column of \code{xmat}
}
\description{
A version of \code{meddiff} to compute median differences in a different format
}
