% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatter}
\alias{scatter}
\title{Make a module-specific scatterplot}
\usage{
scatter(module_code, marks_matrix, student_overall_median, has_groups)
}
\arguments{
\item{module_code}{Character; the length-5 module code}

\item{marks_matrix}{Numeric, the matrix of marks}

\item{student_overall_median}{The row medians of marks_matrix}

\item{has_groups}{Logical. Will be true if both `student_course.csv` and `course_mappings.csv` exist in the working folder.}
}
\value{
A \code{ggplot} object
}
\description{
Make a module-specific scatterplot
}
